makes message containers into a stateless component and moves messaging to views
